requirements.txt
pydantic
fastapi[standard]
uvicorn[standard]
ipykernel
groq
python-dotenv
langchain